# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نبض الحياة API",
        "description": "API للتطبيق الخيري نبض الحياة",
        "contact": {
            "email": "info@nabdhayah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/wallet": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "الحصول على معلومات المحفظة",
                "operationId": "bd5f91f912c7ac1575e8ca7a8e546540",
                "responses": {
                    "200": {
                        "description": "معلومات المحفظة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "balance": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500
                                                },
                                                "total_donations": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1200
                                                },
                                                "donations_count": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wallet/charge": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "شحن المحفظة",
                "operationId": "25fb1d8a7f2cd3f806d33451ea448076",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم إنشاء طلب الشحن بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم إنشاء طلب الشحن بنجاح"
                                        },
                                        "data": {
                                            "properties": {
                                                "payment_url": {
                                                    "type": "string",
                                                    "example": "https://payment.gateway.com/pay/123"
                                                },
                                                "transaction_id": {
                                                    "type": "string",
                                                    "example": "TXN123456"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wallet/transactions": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "الحصول على تاريخ المعاملات",
                "operationId": "3ea6e910cbeabfdd2ca4db13b8d358b2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "رقم الصفحة",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تاريخ المعاملات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 100
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "charge"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "completed"
                                                    },
                                                    "transaction_id": {
                                                        "type": "string",
                                                        "example": "TXN123456"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-01-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}